sửa giỏ hàng

+++ Audio_TripleV/src/main/java/com/example/demo/repository/GioHangChiTietRepository.java
@@ -16,7 +16,7 @@ public interface GioHangChiTietRepository extends JpaRepository<GioHangChiTiet, 
     Double calculateTotalPrice(@Param("gioHangId") Integer gioHangId);
 
     Optional<GioHangChiTiet> findByGioHangAndSanPhamChiTiet(GioHang gioHang, SanPhamChiTiet sanPhamChiTiet);
-    Optional<GioHangChiTiet> findBySanPhamChiTiet_Id(Long productId);
+    Optional<GioHangChiTiet> findBySanPhamChiTiet_Id(Integer productId);
 
 
 }
